requirements.txt
gradio>=4.44.0
selenium==4.15.0
requests==2.31.0
openai>=1.54.0
python-dotenv==1.0.0
pandas==2.0.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.2.0
google-api-python-client==2.100.0
Pillow==10.0.0
flask==3.0.0
flask-cors==4.0.0
